Safari-609.1.11.1/LayoutTests/fast/css/webkit-mask-crash-implicit.html

<!DOCTYPE html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<style>
.box {
-webkit-mask-repeat: repeat-x; -webkit-mask-repeat-y: inherit;
}
</style>
<script>
if (window.internals)
    testRunner.dumpAsText();
shouldBeEqualToString("document.styleSheets[1].rules[0].style.cssText", "-webkit-mask-repeat-x: repeat; -webkit-mask-repeat-y: inherit;");
shouldBeEqualToString("document.styleSheets[1].rules[0].style.getPropertyValue('-webkit-mask')", "");
shouldBeEqualToString("document.styleSheets[1].rules[0].style.getPropertyValue('-webkit-mask-repeat')", "inherit");
</script>
</head>
<body>
</body>
</html>
